Removed old build scripts and updated readme
