+++ README.md
@@ -1,6 +1,19 @@
-#personal-homepage
-Project to build a personal homepage using
+# Personal Homepage Project
+Building and iterating on a personal site where I can test concepts and things I've learned in programming.
 
-Updated Instructions:
-To make changes to pages, make them in either /content/*.html  or /template/template/html
-Run build3.py to combine /content + /template into full pages that live in /docs.
+### Current tech stack:
+- HTML
+- CSS
+- Bootstrap
+- Python
+
+Python is being used to automate the generation of the site by combining content pages with a template.
+
+### Instructions on updating the website:
+1. To update body content of specific pages, update the pages in /content/: 
+2. To update page specific values like `title` &  `meta-description`, update the `pages` dict contained in `build.py`
+3. To update the common header and footer, update template.html in /templates/
+4. Run build.py to automtically apply changes to all user facing "full pages" in /docs/
+```sh
+$ python3 build.py
+```
