+++ build.py
@@ -1,31 +1,64 @@
-#Define template
-top = open('./templates/top.html').read()
-bottom = open('./templates/bottom.html').read()
+#list of every page and the strings to replace
+pages = [ {
+        "filename": "content/index.html",
+        "output": "docs/index.html",
+        "title": "Personal Projects & More",
+        "meta_description": "Welcome to Reuben's personal project site",
+        "index_class": 'class="current-item"',
+}, 
+{
+        "filename": "content/blog.html",
+        "output": "docs/blog.html",
+        "title": "Blog",
+        "meta_description": "Blogging about stuff",
+        "blog_class": 'class="current-item"',
+}, 
+{
+        "filename": "content/contact.html",
+        "output": "docs/contact.html",
+        "title": "Contact Me",
+        "meta_description": "Contact me through the contact form on this page",
+        "contact_class": 'class="current-item"',
+},
+{
+        "filename": "content/projects.html",
+        "output": "docs/projects.html",
+        "title": "Projects & More",
+        "meta_description": "Current & Past Coding Projects",
+        "projects_class": 'class="current-item"',
+},
+]
 
-#Define content for index page
-content = open('./content/index.html').read()
+def replace_template_strings(page):
+    """replace strings in template"""
+    #using Template strings and safesubstitute to replace values
+    #moved the import into this function as it is only used by this function
+    from string import Template
+    template_text = open('./templates/template.html').read()
+    template = Template(template_text)
+    print('Updating:', page['filename'])
+    updated_page = template.safe_substitute(page)
+    #writing the updates to the/doc/pages
+    open(page['output'], 'w+').write(updated_page)
 
-#Define combined page
-index_page = top + content + bottom
+def replace_content_section(page):
+    """replace the content section with /content/ pages"""
+    content = open(page['filename']).read()
+    doc_page = open(page['output']).read()
+    #replacing and writing the doc pages with content
+    full_page = doc_page.replace("{content}", content)
+    open(page['output'], 'w+').write(full_page)
+    return print('Completed: ',page['output'])
 
-#Build combined pages into /docs
-open('./docs/index.html', 'w+').write(index_page)
+def main():
+    i = 0
+    for page in pages:
+        replace_template_strings(page)
+        replace_content_section(page)
+        i += 1
+    #using i to count how many times the function looped 
+    #and comparing with total number of list items
+    return print('- - - -',i,"out of", len(pages), "html pages in /doc/ updated. - - - -")
 
-#Test if it works for the index page
-print('Successfully built the index.html page')
-
-#Repeat above for blog, projects & contact
-content = open('./content/blog.html').read()
-blog_page = top + content + bottom
-open('./docs/blog.html', 'w+').write(blog_page)
-print('Successfully built the blog.html page')
-
-content = open('./content/projects.html').read()
-projects_page = top + content + bottom
-open('./docs/projects.html', 'w+').write(projects_page)
-print('Successfully built the projects.html page')
-
-content = open('./content/contact.html').read()
-contact_page = top + content + bottom
-open('./docs/contact.html', 'w+').write(contact_page)
-print('Successfully built the contact.html page')
+if __name__ == "__main__":
+    main()
